action: Add common "run_at" argument to influence order
One initialconditions/renewal difference is when it runs. To share code
we'll need to make the position in order-of-calculations configurable.

This is (probably) a useful feature to expose generally, and makes the
position in the actions more obvious. So do the same for all actions.

Simplify time so we don't have to have multiple run_at values.

Add a step_id helper for forming the strings.