stock: Allow uneven stock structure
To support likelihood components, uneven stock structures will be
needed. Change g3_stock to take a vector of minlengths, previous
behaviour can be achieved with seq()

As a result, __dl becomes a vector. __countlen is useless so remove it.

Assume internally that the plus group is as big as the mean of the
length groups. This is a bit of a ropey assumption, but should be
practically irrelevant and saves having to handle Inf.