stock_time: Ignore anything outside our range
To be consistent with other intersect methods, anything ouside our
range should be filtered, i.e. code within won't run.

For likelihood actions, this will result in these years getting
ignored (in a rather expensive way, as all the predictions are still
collected). I'm not sure this is the desired end result though.