run: Hide g3_collate, do it implicitly
Instead of having g3_collate exposed, do this as part of compiling the
actions, and just manage lists of actions externally.

This makes it easier to manage replacing of actions, since you don't
have a mess of combined actions to tidy up if you want to e.g. replace
a likelihood component.