action_age: Where possible, just move age groups
We only need to sum 2 age groups when dealing with a plus-group, the
rest of the time values can be wholesale moved, as long as we have an
extra special case for the bottom as well as the top.

This saves work so speeds things up, and makes results align with
gadget2 more.