action_grow: Force only one output_stock, only mature if part of it
Fish only mature if there's somewhere for them to mature *to*,
currently if they don't get picked up by any of the output_stocks (e.g.
because they're in the wrong age group), they get thrown away.

As an interim measure, intersect with the output stock so we only
generate a maturation ratio when there's somewhere for them to go.

Long term, we should be putting fish that aren't picked up by any
output_stock back.