action_grow: Only recalculate growth when we need to
The uses of lgamma are expensive, and this isn't getting automatically
hoisted by GCC (which won't help R anyway).

For now at least, statically analyse the growth formula, and for the
"normal" case of only depending on step size, don't recalcualate growth
until it's needed (possibly only once).

Ideally this would be redundant, but easy to rip out at a later date.