action_grow: Use __plusdl to convert length to lengthgroups #13
For the current formulation to work length groups need to be
basically equal. stock__dl is the difference for each lengthgroup, so
if there's a large plus or minus group this is going to produce
nonsense as fish can't grow out of them.

stock__plusdl is the mode of dl, so whilst we still don't consider
uneven length groups properly, we at least do something consistent.