Add g3_report_all() mechanism, use this by default.
Formulas can't explictly report values since they generally don't have
a place to hang g3_report(), which is a bit of a hack anyway since it
should happen once at the end of the model.

Instead, report anything that gets assigned to in the model
automatically. This will exclude parameters and data (since they're
constant through the model).