rub_tmb: Remove NA check for g3_tmb_lower/upper #59
To implement g3_tmb_parscale() we need to know which values are NA and
auto-populate them.

It's been there since the beginning. I suspect that it was an early
attempt of to hide values that weren't being optimised, and we're now
better at doing this.

Regardless, it's a bit nonsensical, since hiding values from
lower/upper would mean that the structure of par & lower/upper no loger
match, making it useless for any practical purposes.