likelihood_distribution: Attempt at oneStepPredict support #63
Modify distribution functions to include a keep, which gets plumbed in
as a DATA_ARRAY_INDICATOR into our model.

This should in theory work, however:-

* We need to be able to off/on the keep support, otherwise it clogs up
models with unnecessary arrays of 1.

* TMB::oneStepPredict seemingly has no concept of array stride,
without I'm not sure how we might off/on e.g. all length observations
for a single timestep. Without we're not convinced this is very useful.