likelihood_data: stock_re dimension should be called "stock_re"
Use the correct name for stock_re, otherwise it doesn't get included in
the merge() operation, resulting in each number being repeated in the
observation array.

Use [['']] in tests so we don't inavertently rely on R's fuzzy column
matching, which hid this mistake.

In addition, add a global test to make sure in all cases we have the
numbers we expect in the observation array (fixing an unrelated case
which was broken).