renewal: Break normalparam down into input formulas
g3a_initialconditions_normalparam() is pretty old, and a lot of
facilities to make doing this weren't there at the time. But now it's a
lot easier.

This means we have one implementation of both the normal distribution
and the stock-shuffling code to renew.

g3a_initialconditions_normalparam should be deprecated in preference
for g3a_renewal_len_dnorm eventually.