likelihood_distribution: Shouldn't vectorise when over = length
Stratified sumofsquares is g3l_distribution_sumofsquares(over =
c('area', 'length')), which we need to support here.

Do so by ignoring vectorisation and calculating each length in turn.

Dropping the final sum() as we do is a terrible hack, but without we
can't run under TMB. This needs a better solution.