stock: Break up iterate into list of calls
We want to be able to off/on individual parts of iterating over a
stock, to help achieve this, don't combine as we go, keep a list to
combine later in step.R.

Keep the previous iterate-over-everything-but-length behaviour
hard-coded in repl_stock_fn for now. But it's possible to change
finally.

Move the environment out into it's own property, this:
* Removes another copy_all_env user
* Makes finding variables much more obvious
* Makes replacing variables simpler