likelihood_distribution: Use rowSums() to calculate stratified totals
Instead of calculating the total every single time, put it in a
variable. This gives a bit of saving off the bat.

Next, implement stratified sums-of-squares by rowSums instead of
transposing every time we need a lengthgroup. Store this in our total
and fish out the relevant value when required. This should be a
dramatic speed-boost.

Hard-code the __length_idx for now, but ideally we'd have a better way
of getting this.