stock: Expose stock_definition as g3_stock_def #84
Make a new g3_stock_def() that's public, and use that instead of
gadget3:::stock_definition, which we keep around but warn that you
should update.

Whilst here, remove explicit "stock__" from calls. It's unnecessary and
will get in the way when re-factoring the mess of naming and renaming
stocks in g3_step(). Stocks should be using their own name for
variables.