likelihood_distribution: Add transform_fs option #40
Add a hook for adding in a transform after collating data from the
model. The interface is a bit basic, but the intention isn't for it to
be used directly. Eventually we'll have something to hook into here to
build the matrix for us.

We also only support transforming age for now, although the
machinations are built-up for any dimension. Ideally we'd just do
stock_interact(), but that's not designed for interacting a stock with
itself.