timedata: Allow age columns #22
Be more flexible in what we accept, allow year to be missing & include
age in the list of columns that we understand, in which case no lookup
is necessary.

This is a bit dicey whilst sticking with the existing combined-int
method of lookup; we don't really know what the maximum age/area is
going to be, there's potentially many more than available in the table,
and integer overflow is also a real risk.

However, real-world usages should mean being bitten by this is fairly
unlikely.