NEWS: Keep a title in development versions
For the documentation site to build, we have to have a valid news file.
For a file to be valid, we need an initial header.

Add a -999 header on switch to development version, strip it off before
deciding if we need to label news for this version.

Add development header to existing NEWS entries, so it becomes valid
once more.