run_tmb: Explicitly choose beween vector & array
The differing facilities of vector<Type> aren't actually used anywhere,
and the potential for bugs sneaking in if types change depending on the
number of dimensions used is high. Especially bad is .col() is a no-op
on vector<Type>.

Remove heuristics from type logic, making C++ type match R code, adding
force_vector() for when we need to make sure something is vector no
scalar. Use this in place of as.array() where necessary.

Set vector names aside & restore in objective function as we do dimension
names.