WIP: Remove subset->scalar clause, but at what cost? #75
Remove auto-scalar-ing from run_tmb, deal with some of the fallout.

Expressions like:

 x[1,2,3] <- y

...are problematic. y could either be a scalar (as with
nllstock__weight) or a vector (as with assigning to reports). So
automatically doing x[1,2,3][[1]] won't do the right thing in all
cases, and having to add it manually (as is being done below) is ugly
to say the least.

Ideally tmbutils::array would sort this out for us.

There's also transform matrix tests failing, which may or may not be
related.