naturalmortality: Default mortality_f #113
There's only one option, so make it the default.

Don't bubble up the by_stock / by_age options, as using these is
currently a bit hairy at the moment, see:

  https://github.com/gadget-framework/gadget3/issues/113

If you want a more parameterized M, it's a bit more sensible to do this
yourself, example added to show how to do this.