renewal: Default for g3a_renewal_normalparam(run_f) #118
The TRUE default is never useful, will at least want:
* Only renew at the minimum age
* Not a projection year (in which case there's no data)
And possibly want:
* Step in which to renew at

Break up the default clauses into their own parameters, and combine
them by default.

NB: This is assuming that a stock has an age dimension, which it may
not do. Should we be wrapping in if (stock_has_an_age_dimension)?