renewal: Default for g3a_renewal_normalparam(run_f)
The TRUE default is never useful, will at least want:
* Only renew at the minimum age
* Not a projection year (in which case there's no data)
And possibly want:
* Step in which to renew at

Some easier customisation seems likely to be useful, but for now, add
the common case to run_f.

NB: This is assuming that a stock has an age dimension, which it may
not do. Should we be wrapping in if (stock_has_an_age_dimension)?