step: Replace stock_param_* with stock_prepend() #61 #81
They're basically the same thing, and g3_parameterized() means we don't
care about being as succinct as we could be, so replace both with a
wrapper function that will do the same. Unlike g3_param_table they were
only used internally, so can ditch the previous forms.

This is one less thing to consider when restructuring g3_param_table.

It also allows us to do multiple stock_prepends, so we can have a
by_fleet="" option.