likelihood_random: Add weight parameters, hard-code dnorm sense
As with other likelihood parameters, we should be exposing a parameter
by default.

Using the weight to also flip the sense of dnorm(log = TRUE)
unnecessarily complicates reweighting, so move the reweighting so it
happens separately.

As we assume log_f is logical, test for it. Ideally we'd rename the
parameter also.