Replace g3_report_all with internal g3a_report_vars() #131
Rip out g3a_report_all, and instead:

* Hard-code cur_time increment before any actions
* Move time action after any initialconditions/reporting
* Make a new g3a_report_vars() that does the job that g3_report_all()
did, but in a separate action

It's not yet possible to swap out g3a_report_vars() with anything else,
but this clears the way for doing so.

This means cur_time is an enforced part of a model, but that's hardly a
problem.