likelihood_data: Time dimension at start for surveyindices #142
When computing surveyindices we need a timestep vector for each
age/area/whatever grouping we're making. For this to work under TMB we
need the time dimension to be near the beginning.

Previously, this sorta-worked by always ordering time then area. This
meant surveyindices could group by area, and the dimensions are in a
reasonable order for anything else. However, if doing surveyindices by
anything else, this doesn't work.

Instead, allow the position of the time dimension to be configurable,
and shunt it to the start when we see surveyindices being used.