run_tmb: Support dynamic dims in TMB models #131
For TMB output to be used by g3_fit/gadgetplots, TMB has to support
dynamic dims.

Include a function in the time action to generate these, based on the
generated time action itself. Use this to generate the dynamic dims,
either by including it in the R model or processing them at
g3_tmb_adfun() time (when time-based parameters should be fixed).

Given this has been the only reason to use dynamic dims, the support
should probably be removed and replaced with less flexible options, but
this works for now.