likelihood_bounds: Run-time bounds checking #131
Baking bounds into a model is awkward, and likely to result in mishap.

Instead, add g3_param_lower / g3_param_upper pseudo-functions, which
pull the bounds from the template via. DATA parameters.

As we only update these when generating a new objective function, we
don't have to worry about re-taping.

R has no bounds available to it, so set these to NA.

We also need g3_param_nodef, to stop these actions being used as the
definition for the parameter (which may include defaults, which we
don't).