run_tmb: eval() seq argument before comparing
A sys.call() is returning an unevaluated expression, so comparing it
only works if the value is positive, as identical(quote(1), 1).

This is soon due to be an error. From R-devel@r-project.org:

Comparison operators == and != can be used on language objects
(i.e. call objects and symbols). The == operator in particular often
seems to be used as a shorthand for calling identical(). The current
implementation involves comparing deparsed calls as strings. This has
a number of drawbacks and we would like to transition to a more robust
and efficient implementation. As a first step, R-devel will soon be
modified to signal an error when the ordered comparison operators <,
<=, >, >= are used on language objects. A small number of CRAN and
BIOC packages will fail after this change. If you want to check your
packages or code before the change is committed you can run the
current R-devel with the environment variable setting

    _R_COMPARE_LANG_OBJECTS=eqonly

where using such a comparison now produces

    >  quote(x + y) > 1
     Error in quote(x + y) > 1 :
       comparison (>) is not possible for language types