action_predate: Rework to use predprey__cons #29
stock__predby_predstock worked up until now because a fleet has no real
dimensionality. Other predators will.

Use g3s_stock_product() to make predprey__cons, a combination of
predator/prey dimensions, and rework actions to use this.

Store the suitable stock in predprey__suit, whilst this is a bit of a
waste of memory, the alternative would be to calculate suitable stock
twice, which is a waste of tape.

Generate stock__predby_predstock from predprey__cons for
backwards-compatibility, as tagging, likelhood and g3_fit will all have
to be rewritten before it can be removed.

Rework test-action_predate-catchability.R to be an actually useful
test, proving we hit the quotas set for us. Previously the linearfleet
numbers were too high, and we were just overconsuming.