timedata: Simplify intlookup variable names #154
As part of #154, we'll need to add a prefix to a stock dimension. To
simplify this process, rework intlookup naming so the top-level global
formula just uses the passed-in name, and add keys and values to that.

Rename the stock intlookups so they all follow the same pattern,
keeping the stock prefix untouched.