likelihood_data: Do repl_env_fn regardless of prefix #154
repl_env_fn(), as well as adding the predator_ prefix, would modify the
stock name if already set, mostly for the __agegroup
g3_global_formula().

This last part is just as essential without a prefix, so remove the if
conditions always run the replacements. We could trim the replacement
list further, but if we ever get rid of stock__ this will become
important.