action_predate: Remove pred_area from __cons #29
It's a needless waste of memory and backward compatibility problems,
since there will never be consumption across areas.

Instead, rewrite total_predsuit to predstock__totalsuit, basically as
it was before with __catch. With this, we can strip the pred_area
dimension from predprey__cons without ill effect.

Remove the extra dimension we added to all the tests.