action_time: Move cur_year/etc accounting to start of model #155
A g3a_otherfood may well be parameterised by_year. For this to work we
need to have set year in the first place.

The hard-coded cur_year++ was always a bit of a hack, instead make
action_time generate 2 steps, one to do the bulk of the accounting at
the very start of the model, one to stop the model after
initialconditions/reporting have happened.