suitability: Cast andersen suitability to vector<Type> #161
When using the TMBad framework, tests/test-suitability-report.R fails
when trying to use g3_suitability_andersen() for a predator, not a
fleet.

Reverting 1885e4095dec06e62274b262cf3929baf2be4f09 also solves this,
but this isn't something we'd do ideally.

Add a ugly g3_cast_vector() to get the appropriate cast in the right
place. Ideally it wouldn't be necessary, but it's relatively isolated
and could be removed later.