stock: Add stock_isdefined(), use to decide to intersect/iterate
stock_intersect assumes dimensions from the intersecting stock are
present in the outer stock. Generally this is the case but obviously
not generally true.

For individual data points it most likely isn't (e.g. there's no age
data), and we should iterate over all possible values in this case.

When doing stock_intersect, generate both possible bits of code, and
use stock_isdefined() to decide which is relevant after the fact.

This invalidates some tests in test-likelihood_data, as they don't set
up a stock to intersect with. It's not actually a problem with
predation.