likelihood_sparsesample: predstocks option to compare catch data
Add predstocks option, which allows comparison of catch data instead of
abundance.

In the process, pick apart measurement to measurement & quantity, where
the former is the measurement (read: blubber thickness) for the current
area/age/length/... and quantity is either abundance / catch. This
moves the weighting by quanty outside the square in squared-sum, which
was a bug before.

Use ifelse(finites, weighting, 0) in g3l_sparsesample_sumsquares, so
when there is no data we don't return a zero-length array in R.

Rework docs based around current variable names, as well as including
catch options.