run_r: Don't attach globalenv() to models
Using globalenv() has potential for all sorts of weird
action-at-a-distance. Instead, attach to one-above globalenv(), so we
include all attached packages but not globalenv itself.

This breaks finding the model_fn for default parameters, which was
probably a bit dubious anyway. Instead, explicitly add it to the model
environment and use that.