params: Include selection tuple in g3_param_table() call
We use static code analysis to:

* Find out which loop to put code in
* Decide when to run selectivity functions

...both of which fail on the odd special case of g3_param_table(),
which doesn't make direct reference to the variables it uses.

Fix this by including a select parameter, with the tuple we should be
selecting. We don't use this when constructing the table, but means
that all_dependent_vars() will find the variables.

Remove the age-only hack, since that's no longer necessary.

We could use it of course, but the aim has always to refactor
g3_param_table() anyway.