quota: Replace init_val with interim_value
init_val can only accept numeric input and cannot be a parameter e.g.
It also only has any bearing on the result in specific circumstances
(i.e. before an assessment has run on a model with no pre-projection
time period).

Remove it, and replace with an "interim_value" option, which allows you
to set a value to use in the intervening steps between the end of
landings data and before projected quota values kick in.