param_project: Don't compare to 0 in g3_param_project_*ar1 nll
In the first step of the nll for ar1 methods, we were comparing the
first value to 0. This makes sense in the rwalk method, where you're
comparing differences, but not here.

Using this as part of the nll was forcing recruitment to tend to zero.

Instead, skip the first value when calculating nll.