param_project: Account for loglevel in nll, minimum stddev
The loglevel wasn't accounted for in the nll calculations, causing the
AR1 process to fall to zero. This is easy enough to account for,
provided the loglevel parameter isn't set to -1 (read: last nth value).

With some more work, the value could be found (this happens at the end
of the model, so all values are known), but I suspect it'd be a
nonsensical feedback loop.

Also, remove the restriction of nll requiring stddev > 0 by setting a
minimum stddev. The optimiser will require a continuous function even
if noise isn't wanted, so dnorm() is still useful.