action_grow: Don't produce infinte growth when weight zero #223
When a stock is predated to extinction, it's weight will become zero.
In g3a_grow_apply(), we convert mean weight to total biomass. Without
any individuals, this will be zero & thus the mean weight will be zero.

Dividing by zero here produces infinte growth, which results in the
model becoming NaN. avoid_zero() to duck this.

Arguably a mean weight of zero is nonsense, and we shouldn't do that.
But we currently assume it's a possiblity in several places so this is
at least in keeping with the status quo.