tests: Use appropriately-sized default for g3_param_vector()
If our default is 0, then the model can't run because it's invariably
trying to make an array lookup.

This has been a non-issue up until now, but a g3_tmb_fn() will create a
model starting with the initial parameters. Besides, it's not
unreasonable to want the default parameters to work.