run_tmb: Add type="LOG" to parameter_template
Logarithmic parameters (and their bounds) are transparently converted
to logspace for the model, and back again when relisting.

This will be used for g3_param_project functions, where we want
parameters to be in log-space, and generally potentially useful for
parameters that should be positive, instead of relying on things like
avoid_zero & bounds.