param_project: Use "logarithmic" for mean/stddev params for log-dists
Instead of having "lparam" versions of parameters that users have to
manually log(), use logarithmic so this translation to/from happens
automatically.

As well as being more conistent with the rest of gadget, this means the
naming is more consistent. If you're handing in values already in
logspace, then it's just an ar(1) process.