run_tmb: g3_tmb_fn() shouldn't relist already-listed parameters
unlist()ing instead of g3_tmb_par()ing meant we can double-expoentiate
logarithmic values. Instead, restructure so a already-relisted list is
the structure we're aiming for, so we only relist a vector.

Whilst here, rename incoming parameters to have more useful names,
"par" implies a TMB parameter vector, which isn't necessarily what we
want.

Ideally in the table case we should also be updating lower/upper, but
they're unlikely to be useful.